moved metadata to columns directory
